refactor: 상수 패키지 이름 수정
constants에서 constant로 수정
