+++ src/main/java/racingcar/constant/ErrorMessages.java
@@ -1,4 +1,4 @@
-package racingcar.constants;
+package racingcar.constant;
 
 public class ErrorMessages {
     public static final String BLANK_INPUT = "입력된 값이 없습니다.";
