+++ src/main/java/racingcar/constant/Messages.java
@@ -1,4 +1,4 @@
-package racingcar.constants;
+package racingcar.constant;
 
 public class Messages {
     public static final String INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
